% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wCorr-package.R
\docType{package}
\name{wCorr-package}
\alias{wCorr-package}
\title{Computes weighted or unweighted Pearson, Spearman, polyserial, and polychoric correlations coefficients.}
\description{
See \code{\link{weightedCorr}} for examples.
}
