% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightedCorr.R
\name{weightedCorr}
\alias{weightedCorr}
\title{Run a bivariate correlation on an edsurvey.data.frame}
\usage{
weightedCorr(x, y, method = c("Pearson", "Spearman", "Polyserial",
  "Polychoric"), weights = rep(1, length(x)))
}
\arguments{
\item{x}{a numeric (or numeric factor in case of polychoric and polyserial) vector or an object can be forced to a numeric or factor vector}

\item{y}{a numeric vector (or numeric factor in case of polychoric) or an object can be forced to a numeric or factor vector}

\item{method}{a character string indicating which correlation coefficient (or covariance) is to be computed. One of "Pearson" (default), "Spearman", "Polychoric", or "Polyserial".}

\item{weights}{a numeric vector of weights}
}
\value{
Computed value of the correlation
}
\description{
Implements weighted polyserial, weighted polychoric, weighted spearman, and weighted pearson  correlations
}
\details{
In case of polyserial, x must be the observed ordinal variable, and y the observed continuous variable.
the correlation methods are calculated as described in seperate documentaiton.
}
\seealso{
\ifelse{latex}{\code{cor}}{\code{\link[stats]{cor}}}
}

