\name{wCorr-package}
\alias{wCorr-package}
\alias{wCorr}
\docType{package}
\title{
\packageTitle{wCorr}
}
\description{
\packageDescription{wCorr}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{wCorr}
\packageIndices{wCorr}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{wCorr}

Maintainer: \packageMaintainer{wCorr}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
